+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebCheckoutClient.kt
@@ -135,7 +135,7 @@ class PayPalWebCheckoutClient internal constructor(
     }
 
     /**
-     * Call this method at the end of the web checkout flow
+     * Call this method at the end of the web checkout flow to clear out all observers and listeners
      */
     fun removeObservers() {
         activityReference.get()?.let { it.lifecycle.removeObserver(observer) }